requirements.txt
openai>=1.12.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tenacity>=8.2.0
python-dotenv>=1.0.0

